break token balances card into own file
